Implement q_delete_mid
Delete the middle node in the queue. Use two pointers, slow and fast.
In each round, slow moves one step while fast moves two steps,
so we can find the middle node when the loop is finished.
If there is only one node in the queue, it will be deleted.

Change-Id: Ibd6b5bf84dc7d8abc45dadd0462e05159b2bed7d