Implement q_ascend and q_descend
Use two pointers to iterate through the list, pointing to two adjacent
nodes in each round. For ascending order, if the value of the previous
node is greater than that of the next node, delete the previous node.
For descending order, if the value of the previous node is smaller than
that of the next node, delete the previous node.
Once completed, return the size of the queue.

Change-Id: Ib4ba6587364207adbb042565673ec50c7625fc48